done oopsies

+++ src/main/java/kea/eksamen/bowling/dto/ProductPriceDTO.java
@@ -1,6 +1,6 @@
 package kea.eksamen.bowling.dto;
 
 
-public record ProductPriceDTO (int productId, double price){
+public record ProductPriceDTO (int productId, int price){
 
 }
